enterprise_automation_framework_final/requirements.txt
# Core Framework Dependencies
pyyaml==6.0
pandas==2.1.4
openpyxl==3.1.2
requests==2.31.0
flask==3.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# UI Automation
playwright==1.44.0
selenium==4.18.0

# Mobile Automation
appium-python-client==3.1.0

# Reporting
allure-pytest==2.13.2
allure-python-commons==2.13.2
jinja2==3.1.2

# LLM Integration
openai==1.3.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
langgraph==0.0.20
google-generativeai==0.3.2
ollama==0.1.7

# RAGAS and Test Generation
ragas==0.1.0
datasets==2.15.0
evaluate==0.4.1

# Database and ORM
sqlalchemy==2.0.23
alembic==1.13.1

# Security and Authentication
passlib==1.7.4
bcrypt==4.1.2
python-jose==3.3.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
tqdm==4.66.1
pillow==10.1.0
opencv-python==4.8.1.78

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development and Code Quality
black==23.11.0
flake8==6.1.0
pylint==3.0.3
mypy==1.7.1

# Optional: Uncomment for additional features
# slack-sdk==3.26.1  # For Slack notifications
# boto3==1.34.0  # For AWS integration
# azure-storage-blob==12.19.0  # For Azure integration
# google-cloud-storage==2.10.0  # For GCP integration